+++ packages/json-rpc/json-rpc-provider-proxy/src/get-sync-provider.ts
@@ -1,43 +1,44 @@
 import type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
 import { getProxy } from "./get-proxy"
-import { AsyncJsonRpcProvider } from "./public-types"
+import { InnerJsonRpcProvider } from "./public-types"
 import { ConnectableJsonRpcConnection } from "./internal-types"
 
+const noop = () => {}
+
 export const getSyncProvider =
-  (input: () => Promise<AsyncJsonRpcProvider>): JsonRpcProvider =>
+  (
+    input: (onResult: (x: InnerJsonRpcProvider | null) => void) => () => void,
+  ): JsonRpcProvider =>
   (onMessage) => {
     let proxy: ConnectableJsonRpcConnection | null = getProxy(onMessage)
 
-    const start = () => {
-      input().then(
-        (cb) => {
-          if (!proxy) {
-            try {
-              cb(
-                () => {},
-                () => {},
-              ).disconnect()
-            } catch (_) {}
-          } else
-            proxy.connect((onMsg, onHalt) =>
-              cb(onMsg, () => {
-                onHalt()
+    let stop = noop
+    let start = () => {
+      const token = setTimeout(() => {
+        stop = input((cb) => {
+          stop = noop
+          if (!cb) {
+            start()
+          } else if (proxy)
+            proxy.connect((onMsg, onHalt) => {
+              return cb(onMsg, (e) => {
+                onHalt(e)
                 start()
-              }),
-            )
-        },
-        () => {
-          proxy && setTimeout(start, 0)
-        },
-      )
+              })
+            })
+        })
+      }, 0)
+      stop = () => clearTimeout(token)
     }
-
     start()
+
     return {
-      send: (msg) => {
+      send(msg) {
         proxy?.send(msg)
       },
-      disconnect: () => {
+      disconnect() {
+        stop()
+        stop = noop
         proxy?.disconnect()
         proxy = null
       },