+++ packages/json-rpc/sm-provider/src/sm-provider.ts
@@ -5,10 +5,13 @@ import {
   JsonRpcDisabledError,
   type Chain,
 } from "@polkadot-api/smoldot"
-import type { JsonRpcProvider } from "@polkadot-api/json-rpc-provider"
+import type {
+  JsonRpcProvider,
+  JsonRpcResponse,
+} from "@polkadot-api/json-rpc-provider"
 import { getSyncProvider } from "@polkadot-api/json-rpc-provider-proxy"
 
-let pending: Promise<Chain> | null
+let pending: Promise<any> | null
 
 const isRecoverable = (error: any) =>
   !(
@@ -18,41 +21,61 @@ const isRecoverable = (error: any) =>
     error instanceof JsonRpcDisabledError
   )
 
-export const getSmProvider = (chain: Chain | Promise<Chain>): JsonRpcProvider =>
-  getSyncProvider(async () => {
-    while (pending) await pending
+const noop = () => {}
+export const getSmProvider = (
+  inputChain: Chain | Promise<Chain>,
+): JsonRpcProvider =>
+  getSyncProvider((onReady) => {
+    let isRunning = true
 
-    let resolvedChain: Chain
-    if (chain instanceof Promise) {
-      pending = chain
-      resolvedChain = await chain
-      pending = null
-    } else resolvedChain = chain
+    const start = (): any => {
+      if (!isRunning) return
+      if (pending) return pending.then(start)
 
-    return (listener, onError) => {
-      let listening = true
-      ;(async () => {
-        do {
-          let message = ""
-          try {
-            message = await resolvedChain.nextJsonRpcResponse()
-          } catch (e) {
-            if (listening && isRecoverable(e)) onError()
-            return
+      const onResolveChain = (chain: Chain) => {
+        let isListening = isRunning
+        if (!isListening) return chain.remove()
+
+        onReady((onMsgOut, _onhalt) => {
+          const onhalt = (e?: any) => {
+            if (isListening) {
+              isListening = false
+              if (isRecoverable(e)) _onhalt(e)
+            }
           }
-          if (!listening) break
-          listener(message)
-        } while (listening)
-      })()
 
-      return {
-        send(msg: string) {
-          resolvedChain.sendJsonRpc(msg)
-        },
-        disconnect() {
-          listening = false
-          resolvedChain.remove()
-        },
+          const onMsgIn = (msg: JsonRpcResponse | null = null) => {
+            if (isListening!) return
+            msg && onMsgOut(msg)
+            chain
+              .nextJsonRpcResponse()
+              .then((x) => onMsgIn(JSON.parse(x)), onhalt)
+          }
+          onMsgIn()
+
+          return {
+            send(msg) {
+              isListening && chain.sendJsonRpc(JSON.stringify(msg))
+            },
+            disconnect() {
+              isListening = false
+              chain.remove()
+            },
+          }
+        })
       }
+
+      if (inputChain instanceof Promise) {
+        pending = inputChain.catch(noop)
+        inputChain.then(onResolveChain, (e) => {
+          if (isRunning && isRecoverable(e)) onReady(null)
+        })
+        pending = null
+      } else onResolveChain(inputChain)
+    }
+    start()
+
+    return () => {
+      isRunning = false
     }
   })