+++ packages/json-rpc/ws-provider/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./types-common"
 export * from "./types-new"
-export { getWsProvider } from "./default-provider"
+export { getWsProvider } from "./provider"