+++ packages/json-rpc/sm-provider/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- BREAKING: it exposes the new `JsonRpcProvider` v1.
+
 ## 0.1.7 to 0.1.12 - 2025-10-09
 
 ### Fixed